Pivot query logic to new schema
Description:
Everything has begun to adapt to the new JSON-focused schema in the
database but this is just an initial checkin for the code. Several
pieces of logic still need to be written (e.g. passing into the worker
handlers, logic in the handlers for open/closed objects, etc).

Note: Code is passing (but I think silently failing).
Note: Formatting has not been applied.
Note: Refactoring is needed on some of this logic.

Build well