Move conflation logic assets
Description:
Conflation needs to be moved to be present with individual repositories
(see the ArchRepo struct) for possible expansions to onboarding logic
(e.g. front-end heuristics); also note that the Timer assets will now
be on the RepoServer so that it can dictate calls to the database (likely
separate from the server struct) while bridging access to needed pieces
of information.

Note: Only minimal unit tests are present here.
Note: Code is incomplete in the server.go file.

Build well