Move to sub-directory organization
Description:
The query logic will actually be in a sub-directory of the onboarder/
directory since that is a realtively clean spot to put it (the logic is
referenced both by 1. the listening server spinning up a new client
repository and 2. periodic timer calls).

Note: This is a temporary checkin.
Note: Code is incomplete.

Build well