Implement pure database logic
Description:
This is the finalized structure I have settled on for this section
of the codebase - databse w/ methods specific to it, a a RepoServer
with "parent methods" like Timer which control periodic action, and
conflation logic references.

I have also tried to slim down this actual function here but some
minor refactoring would likely be in order.

Note: There are no unit tests for this file yet.

Build well