Gut AddModel method
Description:
Most of the code in the AddModel method is legacy and has
been removed; additionally, the unit test associated with
it has been written (although it may is not complete).

Note: Unit tests are passing.
Note: Code coverage may not be complete.
Note: This code may be further refactored down the road.

Build well