Implement database testing mock
Description:
This is essentially an identical copy to the unit tests available
in the database/sql package for Go. There are still several assets
in this file that can be removed (helpers or used in non-used unit
tests). This code does compile but there are no implemented unit
tests for the Heupr code at this time.

Note: Code comiples.
Note: There are no implemented unit tests.

Build well