Implement engine restructuring
Description:
All of the changes here are part of the initial restructuring of the engine with
only two directories - frontend/ and backend/ - left. There are several major
changes that need to be implemented at this point and the code is currently
broken. I'm going to have to build several assets (e.g. mutexes) as well as
make large restructuring changes (e.g. turn functions into methods).

Note: Code is failing/broken.

Build well