Broad frontend refactoring
Description:
Several of the logic chunks of the engine/ have been split into
only two folders - frontend/ and backend/. This is just the frontend
source code. Most of this code is complete but there may be some
minor pieces of refactoring that need to be completed.

Note: Unit tests are all passing.
Note: The code coverage is not comprehensive.

Build well