Broad backend refactoring
Description:
This is the backend part of the engine checkin; this is a
temporary point because there is still code that needs to
be fully implemented (see Timer). However, it is very near
completion and should be able to be added into the backtest.

Note: Unit tests are passing.
Note: The code coverage is not comprehensive.
Note: Code is partially incomplete.

Build well