Build additional initialization method
Description:
An initialization method was necessary for the scenario in which the server is
being launched for the first time and the RetrieveBulk method is called in
order to avoid a panic.

Note: Unit tests are all passing.
Note: A temporary storage_test.db file is built in the unit test.
Note: The subsequent "delete" test only removes one bucket.
Note: This method has a somewhat more comprehensive unit test.
Note: The actual production method call will be applied in a separate commit.

Build well