Implement activation service
Description:
A "middleman" logic piece has been layered on top of the ingestor/ and backend/
directories to send them notices of new users activating the Heupr service.

This logic has a server of its own and is responsible for performing security
checks and pulling out the necessary information from the request body. A
planned future expansion will also include the user token being passed along
via this route (without hitting BoltDB).

Note: Unit tests are passing and relatively comprehensive.
Note: There is a planned expansion to this piece of code.