Reorder ports and refactor servers' methods
Description:
The various ports that the different pipeline servers listen on and handlers
post to have been properly ordered in the 8000s range; these may ultimately
change for whatever reason.

Also, the user tokens are no longer pulled from storage in the activation
process flow and are now passed directly in the post methods.

Initial groundwork has been laid for the supporting methods on the Ingestor
server. This is a semi-placeholder commit as a result.

Note: Refactoring is still needed for Ingestor (and the other servers, likely).
Note: Code is passing unit tests (in all impacted directories).
Note: General cleanup is still needed in the code.