Implement ingestor restart method
Because there may be interruptions to the ingestor server, a "restart" method was needed to update the database with information that may have been missed while the server was offline. Because models will likely eventually account for more than just issues/pulls, this method is subject to change. A note has been made in the IssueListByRepoOptions regarding the use of the "Since" option that could possibly be a fine tuning.

Note: Code is building correctly.
Note: Unit tests have not been written for this file.
Note: Formatting has been applied.