Improving internal server clarity
Description:
Several of the internal URLs that the various
handlers would respond to were inconsistent and
in some cases incorrect. I made them more
uniform and also made some adjustments to some
erroneous FormValue parsing logic as it did not
conform to the correct structure that the
frontend was putting it into.

Note: Several files/functions are still missing
      proper unit tests.